Rename `NewMetric*` to `ExperimentMetric*`
